Move failure chart file location
